requirements.txt
numpy~=1.19.5
torch~=1.13.1+cu116
opencv-python~=4.5.5.64
torchvision~=0.14.1+cu116
pyyaml~=6.0
scikit-image~=0.19.3
lmdb~=1.4.1
tqdm~=4.64.1
requests~=2.28.2
scipy~=1.4.1
h5py~=3.8.0
einops~=0.6.1